img_request_t.py: rearrange some test code
