+++ tests/img_t.py
@@ -15,37 +15,8 @@ except ImportError:  # Python 2
     from urllib import unquote
 
 from loris import img, img_info
+from loris.img_request import ImageRequest
 from tests import loris_t
-
-
-class TestImageRequest(object):
-
-    @pytest.mark.parametrize('args, request_path', [
-        (('id1', 'full', 'full', '0', 'default', 'jpg'),
-         'id1/full/full/0/default.jpg'),
-        (('id2', '100,100,200,200', '200,', '30', 'gray', 'png'),
-         'id2/100,100,200,200/200,/30/gray.png'),
-    ])
-    def test_request_path(self, args, request_path):
-        request = img.ImageRequest(*args)
-
-        assert request.request_path == request_path
-
-    @pytest.mark.parametrize('args, is_canonical', [
-        (('id1', 'full', 'full', '0', 'default', 'jpg'), True),
-        (('id2', 'full', 'full', '30', 'default', 'jpg'), True),
-        (('id3', 'full', '100,20', '30', 'default', 'jpg'), True),
-
-        # This is a best-fit size, so the canonical size parameter is "80,".
-        (('id4', 'full', '!80,20', '30', 'default', 'jpg'), False),
-    ])
-    def test_is_canonical(self, args, is_canonical):
-        info = img_info.ImageInfo(None)
-        info.width = 100
-        info.height = 100
-        request = img.ImageRequest(*args)
-
-        assert request.is_canonical(info) == is_canonical
 
 
 class Test_ImageCache(loris_t.LorisTest):
@@ -92,7 +63,7 @@ class Test_ImageCache(loris_t.LorisTest):
         image_info = img_info.ImageInfo(None)
         image_info.width = 100
         image_info.height = 100
-        image_request = img.ImageRequest(ident, 'full', 'full', '0', 'default', 'jpg')
+        image_request = ImageRequest(ident, 'full', 'full', '0', 'default', 'jpg')
         self.app.img_cache.create_dir_and_return_file_path(image_request, image_info)
         #call request again, so cache directory should already be there
         # throws an exception if we don't handle that existence properly
@@ -100,20 +71,20 @@ class Test_ImageCache(loris_t.LorisTest):
 
     def test_missing_entry_is_keyerror(self):
         cache = img.ImageCache(cache_root='/tmp')
-        request = img.ImageRequest('id1', 'full', 'full', '0', 'default', 'jpg')
+        request = ImageRequest('id1', 'full', 'full', '0', 'default', 'jpg')
 
         with self.assertRaises(KeyError):
             cache[request]
 
     def test_missing_entry_gets_none(self):
         cache = img.ImageCache(cache_root='/tmp')
-        request = img.ImageRequest('id1', 'full', 'full', '0', 'default', 'jpg')
+        request = ImageRequest('id1', 'full', 'full', '0', 'default', 'jpg')
 
         self.assertIsNone(cache.get(request))
 
     def test_getitem_with_unexpected_error_is_raised(self):
         cache = img.ImageCache(cache_root='/tmp')
-        request = img.ImageRequest('id', 'full', 'full', '0', 'default', 'jpg')
+        request = ImageRequest('id', 'full', 'full', '0', 'default', 'jpg')
 
         message = "Exception thrown in img_t.py for Test_ImageCache"
         m = mock.Mock(side_effect=OSError(-1, message))
@@ -125,5 +96,5 @@ class Test_ImageCache(loris_t.LorisTest):
         # Because this operation is a no-op, we just check we can call the
         # __del__ method without an error.
         cache = img.ImageCache(cache_root='/tmp')
-        request = img.ImageRequest('id1', 'full', 'full', '0', 'default', 'jpg')
+        request = ImageRequest('id1', 'full', 'full', '0', 'default', 'jpg')
         del cache[request]
